secretmanager/v1/python/google-cloud-secret-manager/requirements.txt
google-cloud-secret-manager==2.26.0